Refactor manual edit live routes
